ex3_4 ajout des test des classes Employe, Vendeur

+++ src/main/java/fr/uvsq/coo/ex3_4/Employe.java
@@ -3,9 +3,9 @@ package fr.uvsq.coo.ex3_4;
 public class Employe extends Professionnel{
 	
 	protected double salaireDeBase;
-	protected byte anciennete;
+	protected int anciennete;
 	
-	public Employe(double salaireDeBase, byte anciennete){
+	public Employe(double salaireDeBase, int anciennete){
 		this.salaireDeBase = salaireDeBase;
 		this.anciennete = anciennete;
 	}
